llm_eval-master/llm_eval-master/requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anthropic==0.3.2
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
BLEURT @ file:///Users/cloudwise/Documents/workspace/develop_work/LLM/%E5%A4%A7%E6%A8%A1%E5%9E%8B%E8%AF%84%E4%BC%B0/llm_eval/bleurt
boto==2.49.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.6
click==8.1.4
clu==0.0.9
colorama==0.4.6
comm==0.1.3
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
datasets==2.10.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.8.0
dm-tree==0.1.8
docker-pycreds==0.4.0
editdistance==0.6.2
einops==0.3.0
et-xmlfile==1.1.0
etils==1.3.0
evaluate==0.4.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.17.1
ffmpy==0.3.1
filelock==3.12.2
flatbuffers==23.5.26
flax==0.7.0
fonttools==4.41.0
fqdn==1.5.1
frozenlist==1.4.0
fschat==0.2.20
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
gevent==23.7.0
gin-config==0.5.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
gradio==3.37.0
gradio_client==0.3.0
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
immutabledict==2.2.5
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jax==0.4.13
jaxlib==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
linkify-it-py==2.0.2
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.2.0
markdown2==2.4.9
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mesh-tensorflow==0.1.21
mistune==3.0.1
ml-collections==0.1.1
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.1
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nh3==0.2.14
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.3.1
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandas-stubs==2.0.1.230501
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
portalocker==2.7.0
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
pyglove==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.27.1
seqio-nightly==0.0.15.dev20230731
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
t5==0.9.4
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.0
tensorflow-text==2.13.0
tensorstore==0.1.40
termcolor==2.3.0
terminado==0.17.1
tf-slim==1.1.0
tfds-nightly==4.9.0.dev202304110044
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.0
types-pytz==2023.3.0.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.23.1
wandb==0.15.5
wavedrom==2.0.3.post3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.0
zope.event==5.0
zope.interface==6.0
